Refactor API router in to separate files
